Move utility function to download library
